Application/requirements.txt
numpy
pandas
Flask
scikit-learn
torch
torchvision
requests
Pillow
gunicorn == 20.0.4